python-services/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
openai==1.3.7
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.1.4
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
beautifulsoup4==4.12.2
PyPDF2==3.0.1
python-docx==1.1.0
markdown==3.5.1
scikit-learn==1.3.2
celery==5.3.4
aiofiles==23.2.1